migration işlemleri

+++ HerkesYazarOlsun.BLL/Abstract/IBooksStarsService.cs
@@ -8,7 +8,7 @@ namespace HerkesYazarOlsun.BLL.Abstract
         BooksStars? Guncelle(BooksStars usr, long tck);
         BooksStars? Ekle(BooksStars usr, string? mail);
         BooksStars? Get(long LoginUserId, long YazarId);
-        List<BooksStars> GetBooksStarsByuserId(int userId);       
+        List<BooksStars> GetBooksStarsByuserId(long userId);       
 
     }
